# Improve handling undefined payloads