--- a/lib/reply.js
+++ b/lib/reply.js
@@ -37,15 +37,6 @@ Reply.prototype.send = function (payload) {
 
   this.sent = true
 
-  if (payload === undefined && this._isError === false) {
-    this.res.setHeader('Content-Length', '0')
-    if (!this.res.getHeader('Content-Type')) {
-      this.res.setHeader('Content-Type', 'text/plain')
-    }
-    onSendHook(this, '')
-    return
-  }
-
   if (payload instanceof Error || this._isError === true) {
     handleError(this, payload, onSendHook)
     return
@@ -132,8 +123,14 @@ function wrapOnSendEnd (err, payload) {
 }
 
 function onSendEnd (reply, payload) {
+  if (payload === undefined) {
+    reply.sent = true
+    reply.res.end()
+    return
+  }
+
   var contentType = reply.res.getHeader('Content-Type')
-  if (payload && payload._readableState) {
+  if (payload !== null && payload._readableState) {
     if (!contentType) {
       reply.res.setHeader('Content-Type', 'application/octet-stream')
     }